x-dm-connect-chat/requirements.txt
aws-cdk-lib>=2.244.0,<3.0.0
constructs>=10.5.0,<11.0.0
tweepy>=4.14.0
boto3